cf_caller/requirements.txt

functions-framework==3.5.0
requests==2.31.0
google-cloud-pubsub==2.20.0